virtio: support modern devices
fixes #17239 #17238

Reviewed-on: https://review.haiku-os.org/c/haiku/+/6995

+++ src/add-ons/kernel/bus_managers/virtio/VirtioBalloonDevice.cpp
@@ -17,7 +17,7 @@
 
 
 const char*
-get_feature_name(uint32 feature)
+get_feature_name(uint64 feature)
 {
 	switch (feature) {
 		case VIRTIO_BALLOON_F_MUST_TELL_HOST:
