+++ src/add-ons/kernel/bus_managers/virtio/VirtioBalloonPrivate.h
@@ -61,7 +61,7 @@ private:
 			virtio_device*		fVirtioDevice;
 
 			status_t			fStatus;
-			uint32				fFeatures;
+			uint64				fFeatures;
 			::virtio_queue		fVirtioQueues[2];
 
 			physical_entry 		fEntry;
