+++ src/add-ons/kernel/busses/scsi/virtio/VirtioSCSIController.cpp
@@ -16,7 +16,7 @@
 
 
 const char *
-get_feature_name(uint32 feature)
+get_feature_name(uint64 feature)
 {
 	switch (feature) {
 		case VIRTIO_SCSI_F_INOUT:
